requirements.txt
streamlit==1.25.0
openai==0.27.8
toml~=0.10.2
pandas~=2.0.3
tabulate==0.9.0
python-dotenv~=1.0.0
pyperclip~=1.8.2
asyncio~=3.4.3
streamlit_ace
requests~=2.31.0
